Move function toPem to StandardMetadataServiceSpec
