+++ webauthn-server-attestation/src/test/scala/com/yubico/webauthn/attestation/StandardMetadataServiceSpec.scala
@@ -24,6 +24,8 @@
 
 package com.yubico.webauthn.attestation
 
+import java.security.cert.X509Certificate
+import java.util.Base64
 import java.util.Collections
 
 import com.fasterxml.jackson.databind.node.JsonNodeFactory
@@ -60,6 +62,13 @@ class StandardMetadataServiceSpec extends FunSpec with Matchers {
       new SimpleAttestationResolver(metadata, SimpleTrustResolver.fromMetadata(metadata))
     )
   }
+
+  def toPem(cert: X509Certificate): String = (
+    "-----BEGIN CERTIFICATE-----\n"
+      + Base64.getMimeEncoder(64, System.getProperty("line.separator").getBytes("UTF-8"))
+      .encodeToString(cert.getEncoded)
+      + "\n-----END CERTIFICATE-----\n"
+    )
 
   describe("StandardMetadataService") {
 
@@ -102,7 +111,7 @@ class StandardMetadataServiceSpec extends FunSpec with Matchers {
         s"""{
           "identifier": "44c87ead-4455-423e-88eb-9248e0ebe847",
           "version": 1,
-          "trustedCertificates": ["${TestAuthenticator.toPem(caCert).linesIterator.mkString(raw"\n")}"],
+          "trustedCertificates": ["${toPem(caCert).linesIterator.mkString(raw"\n")}"],
           "vendorInfo": {},
           "devices": [
             {
@@ -160,7 +169,7 @@ class StandardMetadataServiceSpec extends FunSpec with Matchers {
           s"""{
           "identifier": "44c87ead-4455-423e-88eb-9248e0ebe847",
           "version": 1,
-          "trustedCertificates": ["${TestAuthenticator.toPem(caCert).linesIterator.mkString(raw"\n")}"],
+          "trustedCertificates": ["${toPem(caCert).linesIterator.mkString(raw"\n")}"],
           "vendorInfo": {},
           "devices": []
         }"""
@@ -198,7 +207,7 @@ class StandardMetadataServiceSpec extends FunSpec with Matchers {
           s"""{
           "identifier": "44c87ead-4455-423e-88eb-9248e0ebe847",
           "version": 1,
-          "trustedCertificates": ["${TestAuthenticator.toPem(cacaca._1).linesIterator.mkString(raw"\n")}"],
+          "trustedCertificates": ["${toPem(cacaca._1).linesIterator.mkString(raw"\n")}"],
           "vendorInfo": {},
           "devices": [
             {
@@ -228,7 +237,7 @@ class StandardMetadataServiceSpec extends FunSpec with Matchers {
           s"""{
           "identifier": "44c87ead-4455-423e-88eb-9248e0ebe847",
           "version": 1,
-          "trustedCertificates": ["${TestAuthenticator.toPem(caCert).linesIterator.mkString(raw"\n")}"],
+          "trustedCertificates": ["${toPem(caCert).linesIterator.mkString(raw"\n")}"],
           "vendorInfo": {},
           "devices": [
             {
